% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.ASCII.file.R
\name{download.ASCII.file}
\alias{download.ASCII.file}
\title{\code{download.ASCII.file} downloads grid data.}
\usage{
download.ASCII.file(url.string, data.type.label, workingFolder, datestring)
}
\value{
A list variable of the file name and succes/failure flag.
}
\description{
This is an internal function. It downloads an ARCMAP ASCII grid file from a URL.
}
\keyword{internal}
