% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCatchmentData.R
\name{extractCatchmentData}
\alias{extractCatchmentData}
\title{Extracts daily catchment mean and variance from netCDF files of the Bureau of Meteorology (Australia) national gridded climate data.}
\usage{
extractCatchmentData(ncdfFilename = file.path(getwd(), "AWAP.nc"),
  ncdfSolarFilename = file.path(getwd(), "AWAP_solar.nc"),
  extractFrom = as.Date("1900-01-01", "\%Y-\%m-\%d"),
  extractTo = as.Date(Sys.Date(), "\%Y-\%m-\%d"), getPrecip = TRUE,
  getTmin = TRUE, getTmax = TRUE, getVprp = TRUE,
  getSolarrad = TRUE, getMortonsPET = TRUE, DEM = "",
  catchments = "", interp_missing_days = T,
  interp_missing_entries = T, interp_abnormal = T)
}
\arguments{
\item{ncdfFilename}{is a full file name (as string) to the netCDF file.}

\item{ncdfSolarFilename}{is the full file name (as string) to the netCDF file.}

\item{extractFrom}{is a date string specifying the start date for data extraction. The default is \code{"1900-1-1"}.}

\item{extractTo}{is a date string specifying the end date for the data extraction. The default is today's date as YYYY-MM-DD.}

\item{getPrecip}{logical variable for extracting precipitation. Default is \code{TRUE}.}

\item{getTmin}{logical variable for extracting Tmin. Default is \code{TRUE}.}

\item{getTmax}{logical variable for extracting Tmax. Default is \code{TRUE}.}

\item{getVprp}{logical variable for extracting vapour pressure. Default is \code{TRUE}.}

\item{getSolarrad}{logical variable for extracting solar radiation. Default is \code{TRUE}.}

\item{getMortonsPET}{logical variable for calculating Morton's potential ET. Note, to calculate set \code{getTmin=T}, \code{getTmax=T},
\code{getVprp=T} and \code{getSolarrad=T}. Default is \code{TRUE}.}

\item{DEM}{is either the full file name to a ESRI ASCII grid (as lat/long and using GDA94) or a raster class grid object. The DEM is used
for the calculation of Morton's PET. The Australian 9 second DEM can be loaded using \code{data(DEM_9s)}. For details see
\url{https://www.data.gov.au/dataset/geodata-9-second-dem-and-d8-digital-elevation-model-version-3-and-flow-direction-grid-2008}}

\item{catchments}{is either the full file name to an ESRI shape file of points or polygons (latter assumed to be catchment boundaries) or a shape file
already imported using readShapeSpatial(). Either way the shape file must be in long/lat (i.e. not projected) use the ellipsoid GRS 80.}

\item{interp_missing_days}{T or F, indicating if missing days should be interpolated for PET calculation. Default is \code{T}. See \code{\link{Evapotranspiration::ReadInputs}}}

\item{interp_missing_entries}{T or F, indicating if missing data entries should be interpolated for PET calculation. Default is \code{T}. See \code{\link{Evapotranspiration::ReadInputs}}}

\item{interp_abnormal}{T or F, indicating if abnormal valuses should be interpolated for PET calculation. Default is \code{T}. See \code{\link{Evapotranspiration::ReadInputs}}}
}
\value{
When "catchments" are polygons, the returned variable is list variables containing two data.frames, one of the catchment average daily climate
metrics and another of the catchment variance daily climate metrics.

When "catchments" are points, the returned variable is a data.frame containing daily climate data at each point.
}
\description{
\code{extractCatchmentData} extracts catchment average climate data from netCDF files containing Australian climate data.
}
\details{
This function extracts the AWAP climate data for each point or polygon, and for the latter the daily spatial mean and variance of
each climate metric are calculated. The calculation of the spatial mean uses the fraction of each AWAP grid cell within the polygon.
The variance calculation does not use the fraction of the grid cell and returns NA if there are <2 grid cells in the catchment boundary.
Prior to the catchment averaging and variance, Morton's areal potential ET (PET) is also calculated; after which the mean and
variance PET is calculated. Morton's PET is calculated using the ET.MortonCRAE() function from the Evapotranspiration package at
a monthly time-step and using the AWAP solar radiation. For both points and polygons, the monthly PET estimate is then interpolated
using a spline to a daily time step (using zoo:na.spline()) and then constrained to >=0. In calculating PET, the input data
is pre-processed using Evapotranspiration::ReadInputs() such that missing days, missing enteries and abnormal values are interpolated
(by default) with the former two interpolated using the "DoY average", i.e. replacement with same day-of-the-year average. Additionally, AWAP solar
radiation data is only available from 1/1/1990. To derive PET values <1990, the average solar radiation for each day of the year from
1/1/990 to "extractTo" is derived (i.e. 365 values) and then applied to each day prior to 1990. Importantly, the estimates of PET <1990
are dependent upon the end date extracted. Re-running the estimation of PET with a later extractTo data will change the estimates of PET
prior to 1990.

Also, when "catchments" is points (not polygons), then the netCDF grids are interpolate using bilinear interpolation of
the closest 4 grid cells.

Lastly, data is extracted for all time points and no temporal infilling is undertaken if the grid cells are blank.
}
\examples{
# Load required packages.
library(sp);library(raster);library(chron);library(ncdf4);
library(maptools);library(Evapotranspiration);library(AWAPer)

# Download the 9 second Australian DEM.
DEM_9s = getDEM()

# Load example cacthment boundaries.
data("catchments")

# Extract all climate data and calculate Morton's PET.
# Note, the input "catchments" can also be a file to a ESRI shape file.
climateData = extractCatchmentData(catchments=catchments,DEM=DEM_9s)

# Extract the daily catchment average data.
climateDataAvg = climateData$catchmentAvg

# Extract the daily catchment variance data.
climateDataVar = climateData$catchmentVar

# Export data to .csv files
write.csv(climateDataAvg,'warrionClimateAvg.csv')
write.csv(climateDataVar,'warrionClimateVar.csv')

}
\seealso{
\code{\link{makeNetCDF_file}} for building the NetCDF files of daily climate data.
}
