% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.ASCII.file.header.R
\name{get.ASCII.file.header}
\alias{get.ASCII.file.header}
\title{\code{get.ASCII.file.header} get header information.}
\usage{
get.ASCII.file.header(des.file.name, data.type.label, workingFolder,
  datestring, remove.file = T)
}
\value{
A list variable of header information.
}
\description{
This is an internal function. It returns the header information within a ARCMAP ASCII grid file.
}
\keyword{internal}
