% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEM.R
\name{getDEM}
\alias{getDEM}
\title{Downloads and imports Geoscience Australia 9s DEM.}
\usage{
getDEM(workingFolder = getwd(),
  urlDEM = "https://s3-ap-southeast-2.amazonaws.com/elvis.ga.gov.au/elevation/9secPackagedData/DEM-9S_ESRI.zip",
  keepFiles = F)
}
\arguments{
\item{workingFolder}{is the file path (as string) in which to download the zip file. The default is \code{getwd()}.}

\item{urlDEM}{URL to the folder containing the Geoscience Australia 9s DEM.
The default is \url{https://s3-ap-southeast-2.amazonaws.com/elvis.ga.gov.au/elevation/9secPackagedData/DEM-9S_ESRI.zip}.}

\item{keepFiles}{is a logical scalar to keep the downloaded zip file and extracted DEM ASCII file. The default is \code{FALSE}.}
}
\value{
A RasterLayer DEM for Asutralia.
}
\description{
\code{getDEM} get Australian 9s DEM.
}
\details{
This function downlaod the Geoscience Australia 9 second DEM and then imports the grid.
The DEM is required for the calculation of evaportranspiration within \code{extractCatchmentData}. For details of the DEM see
\url{https://www.data.gov.au/dataset/geodata-9-second-dem-and-d8-digital-elevation-model-version-3-and-flow-direction-grid-2008}
}
\examples{
# Load required packages.
library(sp);library(raster);library(chron);library(ncdf4);
library(maptools);library(Evapotranspiration);library(AWAPer)

# Download the DEM.
DEM_9s = getDEM()

# Plot the DEM.
image(DEM_9s, xlab='Long.',ylab='Lat.')

# Save DEM for next time it is needed.
save(DEM_9s,file="DEM.RData" )

}
\seealso{
\code{\link{extractCatchmentData}} for extracting catchment daily average and variance data.
}
