% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getURLs.R
\name{getURLs}
\alias{getURLs}
\title{Get default URLs for loading data.}
\usage{
getURLs()
}
\value{
A list variable of URLs as characters.
}
\description{
\code{getURLs} get URLS to AWAP and Australian 9s DEM.
}
\details{
This function returns a list of default URLs used to download the AWAP and DEM data.
}
\examples{

URLs = getURLs()

}
