% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.ASCII.file.R
\name{readin.ASCII.file}
\alias{readin.ASCII.file}
\title{\code{readin.ASCII.file} reads in a grid data.}
\usage{
readin.ASCII.file(file.name, nRows, noData)
}
\value{
A matrix of the grid data.
}
\description{
This is an internal function. It reads in an ARCMAP ASCII grid file.
}
\keyword{internal}
