% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCatchmentData.R
\name{extractCatchmentData}
\alias{extractCatchmentData}
\title{Extracts daily catchment mean and variance from netCDF files of the Bureau of Meteorology (Australia) national gridded climate data.}
\usage{
extractCatchmentData(ncdfFilename = file.path(getwd(), "AWAP.nc"),
  ncdfSolarFilename = file.path(getwd(), "AWAP_solar.nc"),
  extractFrom = as.Date("1900-01-01", "\%Y-\%m-\%d"),
  extractTo = as.Date(Sys.Date(), "\%Y-\%m-\%d"), getPrecip = TRUE,
  getTmin = TRUE, getTmax = TRUE, getVprp = TRUE, getSolarrad = TRUE,
  getMortonsPET = TRUE, DEM = "", catchments = "")
}
\arguments{
\item{ncdfFilename}{is a full file name (as string) to the netCDF file.}

\item{ncdfSolarFilename}{is the full file name (as string) to the netCDF file.}

\item{extractFrom}{is a date string specifying the start date for data extraction. The default is \code{"1900-1-1"}.}

\item{extractTo}{is a date string specifying the end date for the data extraction. The default is today's date as YYYY-MM-DD.}

\item{getPrecip}{logical variable for extracting precipitation. Default is \code{TRUE}.}

\item{getTmin}{logical variable for extracting Tmin. Default is \code{TRUE}.}

\item{getTmax}{logical variable for extracting Tmax. Default is \code{TRUE}.}

\item{getVprp}{logical variable for extracting vapour pressure. Default is \code{TRUE}.}

\item{getSolarrad}{logical variable for extracting solar radiation. Default is \code{TRUE}.}

\item{getMortonsPET}{logical variable for calculating Morton's potential ET. Note, to calculate set \code{getTmin=T}, \code{getTmax=T}, \code{getVprp=T} and \code{getSolarrad=T}. Default is \code{TRUE}.}

\item{DEM}{is either the full file name to a ESRI ASCII grid (as lat/long and using GDA94) or a raster class grid object. The DEM is used for the calculation
of Morton's PET. The Australian 9 second DEM can be loaded using \code{data(DEM_9s)}. For details see 
url{https://www.data.gov.au/dataset/geodata-9-second-dem-and-d8-digital-elevation-model-version-3-and-flow-direction-grid-2008}}

\item{catchments}{is either the full file name to an ESRI shape file of polygons (assumed catchment boundaries) or a shape file
already imported using readShapePoly(). Either way the dhape file must be in long/lat (i.e. not projected) use the ellipsoid GRS 80.}
}
\value{
A list variale containing two data frames of the catchment average daily climate metrics and the catchment variance daily climate metrics.
}
\description{
\code{extractCatchmentData} extracts catchment average climate data from netCDF files containing Australian climate data.
}
\details{
This function extracts the AWAP climate data for each polygon and then calculates the daily spatial average and variance for
each climate metric. The calculation of the mean uses the fraction of each AWAP grid cell within the polygon. The variance 
calculation does not use the fraction of the grid cell.  Prior to the catchment averaging and variance, Morton's poential ET 
is also calculated; average which the mean and variance are calculated.

Lastly, data is extracted for all time points and no temporal infilling is undertaken if the grid cells are blank.
}
\examples{
# Load the 9 second Australian DEM.
data("DEM_9s")

# Load example cacthment boundaries.
data("catchmentBoundaries")

# Extract all climate data and calculate Morton's PET. 
# Note, the input "catchments" can also be a file to a ESRI shape file.
climateData = extractCatchmentData(catchments=catchments,DEM=DEM_9s)

# Extract the daily catchment average data.
climateDataAvg = climateData$catchmentAvg

# Extract the daily catchment variance data.
climateDataVar = climateData$catchmentVar
}
\seealso{
\code{\link{makeNetCDF_file}} for building the NetCDF files of daily climate data.
\code{\link{mortonsAET}} for details of the calculate of areal potential ET.
}
